Extract config
