+++ backtesting/backtesting.py
@@ -1,20 +1,13 @@
 # Libraries
-from datetime import datetime, timedelta
-import typing
 
 from pandas import DataFrame
 from tqdm import tqdm
-import numpy as np
 
 # Files
-from backtesting.results import MainResults, OpenTradeResult, CoinInsights, show_signature
 from backtesting.strategy import Strategy
 from data.tradingmodule import TradingModule
-from modules.setup.config import load_strategy_from_config, ConfigModule
-from modules.setup.config.currencies import get_currency_symbol
-from utils import calculate_worth_of_open_trades, default_empty_dict_dict
-from models.trade import Trade
-from backtesting.plots import plot_per_coin
+from modules.pairs_data import PairsData
+from modules.setup.config import ConfigModule
 
 
 # ======================================================================
@@ -26,7 +19,6 @@ from backtesting.plots import plot_per_coin
 #
 # These constants are used for displaying and
 # emphasizing in commandline backtestresults
-
 
 
 class BackTesting:
@@ -46,25 +38,24 @@ class BackTesting:
         self.backtesting_to = config_module.backtesting_to
         self.btc_marketchange_ratio = config_module.btc_marketchange_ratio
 
-    def start_backtesting(self) -> dict:
+    def start_backtesting(self) -> PairsData:
         print('[INFO] Starting backtest...')
 
         data_dict = self.populate_signals()
         return data_dict
 
-    def populate_signals(self) -> dict:
+    def populate_signals(self) -> PairsData:
         """
         Method used for populating indicators / signals
         Populates indicators
         Populates buy signal
         Populates sell signal
         :return: dict containing OHLCV data per pair
-        :rtype: dict
         """
         data_dict = {}
         notify = False
         notify_reason = ""
-        stoploss_type = self.config['stoploss-type']
+        stoploss_type = self.config.raw_config['stoploss-type']
         for pair in tqdm(self.data.keys(), desc="[INFO] Populating Indicators",
                          total=len(self.data.keys()), ncols=75):
             df = self.data[pair]
@@ -84,7 +75,6 @@ class BackTesting:
                     notify_reason = "configured incorrectly"
             data_dict[pair] = indicators.to_dict('index')
         if notify:
-            print(f"[WARNING] Dynamic stoploss {notify_reason}. Using standard stoploss of {self.config['stoploss']}%.")
+            print(
+                f"[WARNING] Dynamic stoploss {notify_reason}. Using standard stoploss of {self.config.raw_config['stoploss']}%.")
         return data_dict
-
-    # This method is called when backtesting method finished processing all OHLCV-data
