+++ modules/pairs_data.py
@@ -1,0 +1,1 @@
+PairsData = dict[str, dict[str, dict[int, float]]]
